DomainFacade is not responsible for returning DomainConfigs by Currency, Currency facade is now responsible for that
- Domain Component is now independent from framework and is component
