+++ packages/framework/src/Component/Domain/DomainFacade.php
@@ -3,8 +3,6 @@
 namespace Shopsys\FrameworkBundle\Component\Domain;
 
 use Shopsys\FrameworkBundle\Component\FileUpload\FileUpload;
-use Shopsys\FrameworkBundle\Model\Pricing\Currency\Currency;
-use Shopsys\FrameworkBundle\Model\Pricing\PricingSetting;
 use Symfony\Component\Filesystem\Filesystem;
 
 class DomainFacade
@@ -13,11 +11,6 @@ class DomainFacade
      * @var \Shopsys\FrameworkBundle\Component\Domain\Domain
      */
     protected $domain;
-
-    /**
-     * @var \Shopsys\FrameworkBundle\Model\Pricing\PricingSetting
-     */
-    protected $pricingSetting;
 
     /**
      * @var \Shopsys\FrameworkBundle\Component\Domain\DomainService
@@ -42,34 +35,15 @@ class DomainFacade
     public function __construct(
         $domainImagesDirectory,
         Domain $domain,
-        PricingSetting $pricingSetting,
         DomainService $domainService,
         Filesystem $fileSystem,
         FileUpload $fileUpload
     ) {
         $this->domainImagesDirectory = $domainImagesDirectory;
         $this->domain = $domain;
-        $this->pricingSetting = $pricingSetting;
         $this->domainService = $domainService;
         $this->filesystem = $fileSystem;
         $this->fileUpload = $fileUpload;
-    }
-
-    /**
-     * @param \Shopsys\FrameworkBundle\Model\Pricing\Currency\Currency $currency
-     * @return \Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig[]
-     */
-    public function getDomainConfigsByCurrency(Currency $currency)
-    {
-        $domainConfigs = [];
-        foreach ($this->domain->getAll() as $domainConfig) {
-            $domainCurrencyId = $this->pricingSetting->getDomainDefaultCurrencyIdByDomainId($domainConfig->getId());
-            if ($domainCurrencyId === $currency->getId()) {
-                $domainConfigs[] = $domainConfig;
-            }
-        }
-
-        return $domainConfigs;
     }
 
     /**
