+++ packages/framework/src/Model/Pricing/Currency/CurrencyFacade.php
@@ -253,4 +253,21 @@ class CurrencyFacade
 
         $this->em->flush($toFlush);
     }
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Model\Pricing\Currency\Currency $currency
+     * @return \Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig[]
+     */
+    public function getDomainConfigsByCurrency(Currency $currency)
+    {
+        $domainConfigs = [];
+        foreach ($this->domain->getAll() as $domainConfig) {
+            $domainCurrencyId = $this->pricingSetting->getDomainDefaultCurrencyIdByDomainId($domainConfig->getId());
+            if ($domainCurrencyId === $currency->getId()) {
+                $domainConfigs[] = $domainConfig;
+            }
+        }
+
+        return $domainConfigs;
+    }
 }
